changed decimal to integer in migration database

+++ db/migrate/20220124101214_create_appointment.rb
@@ -4,7 +4,7 @@ class CreateAppointment < ActiveRecord::Migration[6.1]
       t.references :service, null: false, foreign_key: true
       t.references :request, null: false, foreign_key: true
       t.datetime :schedule
-      t.decimal :total_charge
+      t.integer :total_charge
       t.datetime :confirmed_at
 
       t.timestamps
