+++ db/migrate/20220124101527_create_review.rb
@@ -3,8 +3,8 @@ class CreateReview < ActiveRecord::Migration[6.1]
     create_table :reviews do |t|
       t.references :appointment, null: false, foreign_key: true
       t.text :client_comment
-      t.decimal :chef_rating
-      t.decimal :client_rating
+      t.integer :chef_rating
+      t.integer :client_rating
 
       t.timestamps
     end
